IPC::Open3::open3() shouldn't fail if any of *STD{IN,OUT,ERR} are localized.
Previously it would fail on Win32, because spawn_with_handles() would attempt to
duplicate all three, ignoring failures at this point, but then report failure to
close any as a fatal error, even if this was because the earlier dup-ing had
failed.

Also avoid a warning in the *nix code path in open3() if STDERR is localized
(and hence fileno STDERR is undefined).