Simplify the test for IPC::Open bug RT #72016.
As the original bug was that errors were reported inconsistently between the
DO_SPAWN/not DO_SPAWN code paths of open3(), the original test could have
expressed that more directly with one code path and conditionally adding a TODO.
The refactoring of bd29e8c290c68f4f failed to spot this, and introduced new
logic errors in the DO_SPAWN path - waitpid() should not be called if $@ is set.

Set $pid outside the eval {} - this makes sure it is (re)set to undef if the
eval fails, instead of holding its previous (now bogus) value.